+++ lib/kubes/auth.rb
@@ -10,11 +10,14 @@ module Kubes
       klass.new(@image).run
     end
 
-    # Currently only support ECR
+    # Currently only support ECR and GCR
+    # TODO: consider moving this to plugin gems
     def strategy_class
       case @image
       when /\.amazonaws\.com/ # IE: 112233445566.dkr.ecr.us-west-2.amazonaws.com/demo/sinatra
         Ecr
+      when /gcr\.io/
+        Gcr
       end
     end
   end