+++ lib/templates/new/variable/file.rb
@@ -1,0 +1,1 @@
+@example = "<%= Kubes.env %>-value"