+++ spec/kubes/dsl/pod_spec.rb
@@ -1,6 +1,6 @@
 describe Kubes::Compiler::Dsl::Syntax::Pod do
   let(:evaluator) { described_class.new(options) }
-  let(:options) { {path: fixture("syntax/pod") } }
+  let(:options) { {path: fixture("syntax/.kubes/resources/web/pod") } }
 
   context "pod" do
     it "run" do